requirements.txt
picamera2
opencv-python
numpy
RPi.GPIO
